Refactor/enhance tests for Memory Repository
