--- tests/provider/elements.py
@@ -1,14 +1,0 @@
-from protean.core.aggregate import BaseAggregate
-from protean.fields import Integer, String
-
-
-class Person(BaseAggregate):
-    first_name = String(max_length=50, required=True)
-    last_name = String(max_length=50, required=True)
-    age = Integer(default=21)
-
-
-class Alien(BaseAggregate):
-    first_name = String(max_length=50, required=True)
-    last_name = String(max_length=50, required=True)
-    age = Integer(default=21)
